Adding first version of the SDCollisionVis plugin
Tested in anger on UE5.6.1 and smoke tested for UE5.7

SDCollisionVis is a plugin for UnrealEngine to help see the actual state
of the scene, as Chaos sees it.
The collision viewmodes that come with Unreal only draw the rendering
mesh, not the physics mesh.
Using `LineTraceSingleByObjectType`, we can raytrace the physics scene
directly and build a better image of what's going on.